Add sv_region to csgoserver
From developer.valvesoftware.com:
"This cvar might be deprecated, it has been removed from the HLDS (but still exists in SRCDS) and the Steam server browser doesn't have an option to select a region any more. (Though the in-game server browser might still have this option.)"

And yes, the in-game server browser still have this option.